requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
openai==1.12.0
httpx==0.25.2
python-dotenv==1.0.0
reportlab==4.0.9
jinja2==3.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
openai>=1.10.0
